requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
flask-cors==4.0.0
python-dotenv==1.0.0
Pillow==10.4.0
requests==2.31.0
opencv-python-headless==4.9.0.80
pytesseract==0.3.10
easyocr==1.7.0
numpy
matplotlib
sympy
moviepy
gtts
openai
setuptools